+++ src/shader/slang/slang_utility.h
@@ -35,7 +35,7 @@
 
 #define slang_string_compare(str1, str2) strcmp (str1, str2)
 #define slang_string_copy(dst, src) strcpy (dst, src)
-#define slang_string_length(str) strlen (str)
+#define slang_string_length(str) ((GLuint) strlen (str))
 
 char *slang_string_concat(char *, const char *);
